pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "canary-mlx"
version = "0.1.0"
description = "An implementation of Nvidia's Canary ASR/AST models for Apple Silicon using MLX."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dacite>=1.9.2",
    "huggingface-hub>=0.30.2",
    "librosa>=0.11.0",
    "mlx>=0.22.1",
    "numpy>=2.2.5",
    "pyyaml>=6.0.2",
    "sentencepiece>=0.2.0",
    "typer>=0.15.3",
]
license = "Apache-2.0"
keywords = [
    "mlx",
    "canary",
    "asr",
    "ast",
    "nvidia",
    "apple",
    "speech",
    "translation",
    "ml",
]

[project.urls]
Repository = "https://github.com/senstella/parakeet-mlx.git"
Issues = "https://github.com/senstella/parakeet-mlx/issues"

[project.scripts]
canary-mlx = "canary_mlx.cli:app"
canary-mlx-diff = "canary_mlx.weight_diff:app"

[project.optional-dependencies]
conversion = [
    "torch>=2.2.0",
]
